Update linux.md
Without d3dcompiler_47 HLSL shader compilation failed. This caused the entire mod manager window to not render at all. Not sure which wine version this is specific to or if there was some change in mod assistant itself at some point. Regardless this seems to have fixed the problem for me.