Adding information about Inverts
Other than in Mapping Glossary there is nothing in the wiki that that discusses inverts. There is a macro in the BSMG explaining it but again, nothing here.
Since I never had any information about them, I was unable to provide proper feedback on them during a testplay the other day that mostly contained inverts. It was only this morning when I was reading though mapping-discussion that I learned what they really are and how they really affected a map.