Reformat table of environments in basic-lighting
Originally, a GFM-style table is used for this table. The problem is
that the table extension is optimized for table entries that fit onto
a single line, which is obviously not the case here. This resulted in
Markdown that is difficult to read in the source code view.

This commit reformats the table by replacing the GFM-style table with
raw HTML. Using raw HTML allows the usage of line breaks within a table,
improving the readability of the Markdown source for our case.