Cover more details of v3 lighting in intermediate
Added a tutorial -- maybe more of a documentation -- of the Group Info
menu. Currently intermediate lighting is where all the stuff about
Group Info lies, but some may be worth moving to basic lighting.

Also documented a few note types that are missing from basic lighting
due to an oversight.