add more detailed info of envs up to The Second
This is a pretty monstrous set of changes in a single commit.

The format of environment description is changed; the full set
of event lanes available for an environment are categorized and
named according to how they look and work, ditching the semi-
semantic classical names applied to Static Event System (SES)
event IDs.

A proper description of Weave, FOB, EDM and The Second is added
using the new description format, complete with a listing of SES
event lanes and Group Lighting System (GLS) objects and notes on
environment objects that require them. Other environments also
receive updated notes, reflecting the changes to them applied by
Beat Games.

The listing of environments up to Panic are reordered to a stricter
chronological order. Not sure how useful this is.

The entry for Glass Desert is left mostly intact, yet. Remaining
environments would be worked on later.