Undocument 360/90 mapping with official editor
As late as 1.27.0, the legacy official editor is no longer accessible.
The new official editor meanwhile does not support 360/90 mapping,
marking the characteristics as "Coming Soon (TM)". As such resources
mentioning the official editor for 360/90 mapping would be difficult to
follow due to the hard requirement of downgrading the game.

As such, it would be desirable to remove the whole section on 360/90
mapping with legacy official editor, as it is now completely irrelevant.
The mention in MMA2's section is left in with some changes; however,
I am finding the claim of ease and effectiveness slightly dubious, as
following downgrade tutorials adds complexity.